Refactor SelectColumnOrder and dependent components
