--- frontend/src/components/SelectColumn/index.ts
@@ -1,1 +1,0 @@
-export * from './SelectColumn';
