+++ frontend/src/components/SelectColumnOrder/StyledItem.tsx
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const StyledItem = styled.div`
+  margin-top: 4px;
+  margin-bottom: 4px;
+  margin-right: 8px;
+  cursor: grab;
+  word-wrap: break-word;
+  line-height: 1em;
+  white-space: normal;
+  padding: 0.78571429em 2.1em 0.78571429em 1em;
+  color: rgba(0, 0, 0, 0.87);
+  border: 1px solid rgba(34, 36, 38, 0.15);
+  border-radius: 0.28571429rem;
+  min-height: 2.71428571em;
+  max-width: 500px;
+  min-width: 300px;
+  width: fit-content;
+`;
+
+export default StyledItem;
